lesson_6_starknet/dz/requirements.txt
aiohttp==3.9.3
aiohttp-proxy==0.1.2
fake-useragent==1.4.0
loguru==0.7.2
python-dotenv==1.0.1
